Move wordlist functionality into MnemonicHelper for reuse by SignUpActivity
